refactor(Datepicker): Refactor package structure
