--- packages/forma-36-react-components-datepicker/src/index.ts
@@ -1,1 +1,0 @@
-export { Datepicker } from './Datepicker/Datepicker';
